+++ index.html
@@ -68,6 +68,8 @@
     </header>
 
     <!-- DOM default display none -->
+    
+
     <menu class="menu menu-visibility">
         <div class="cancel-icon" id="cancel-icon">
             <img src="./assets/IconCancel.svg" alt="">
@@ -78,12 +80,10 @@
             <li class="menuLi"><a href="#" data-target="contact">Contact</a></li>
         </ul>
     </menu>
-
+    <div class="popup-hero" id="cont"></div>
     <!-- Main -->
-    <main>
-        
+    <main id="main">   
         <!-- moved to JS obj -->
-
         <section class="cards content" id="cards">
             <!-- Card One -->
             <section class="card">
@@ -109,7 +109,7 @@
                         <li><a href="#">CSS</a></li>
                         <li><a href="#">JavaScript</a></li>     
                     </ul>
-                    <div class="btn btn-card">See Project</div>
+                    <div class="btn btn-card" id="card-pj1">See Project</div>
             </section>
             <!-- Card Two -->
             <section class="card card-two">
@@ -141,7 +141,7 @@
                         <li><a href="#">CSS</a></li>
                         <li><a href="#">JavaScript</a></li>     
                     </ul>
-                    <div class="btn btn-card">See Project</div>
+                    <div class="btn btn-card" id="card-pj2">See Project</div>
             </section>
             <!-- Card Three -->
             <section class="card">
@@ -174,7 +174,7 @@
                         <li><a href="#">CSS</a></li>
                         <li><a href="#">JavaScript</a></li>     
                     </ul>
-                    <div class="btn btn-card">See Project</div>
+                    <div class="btn btn-card" id="card-pj3">See Project</div>
             </section>
             <!-- Card Four -->
             <section class="card card-four">
@@ -207,7 +207,7 @@
                         <li><a href="#">CSS</a></li>
                         <li><a href="#">JavaScript</a></li>     
                     </ul>
-                    <div class="btn btn-card">See Project</div>
+                    <div class="btn btn-card" id="card-pj4">See Project</div>
                 </div>
             </section>
         </section>
@@ -279,6 +279,7 @@
                 </ul>
             </div>
         </section>
+        
 
         <section class="contactme content" id="contact">
             <div class="contactme-container">
@@ -297,7 +298,8 @@
             </div>
         </section>
     </main> 
-    <script src="./scripts/popup.js"></script>
+    
+    <script src="./scripts/module.js"></script>
     <script src="./scripts/main.js"></script>
 </body>
 </html>
