--- scripts/projects.js
@@ -1,27 +1,0 @@
-const generalArticle = `
-Lorem Ipsum is simply dummy text of the printing and typesetting 
-industry. Lorem Ipsum has been the industry's standard dummy text 
-ever since the 1500s, when an unknown printer took a galley of type 
-and scrambled it to make a type specimen book. It has survived not 
-only five centuries, but also the leap into electronic typesetting, 
-remaining essent
-`;
-
-const projects = [
-  {
-    name: 'Tonic',
-    article: generalArticle,
-    image: './assets/cards/card1.svg',
-    liveLink: '#',
-    sourceLink: '#',
-  },
-  {
-    name: 'Multi-Post Stories',
-    article: generalArticle,
-    image: './assets/cards/card1.svg',
-    liveLink: '#',
-    sourceLink: '#',
-  },
-];
-
-module.exports = projects;
