+++ stylesheets/M_cards.css
@@ -21,6 +21,13 @@
   margin: 0 auto;
 }
 
+.popup-hero {
+  position: absolute;
+  top: 0%;
+  left: 0%;
+  display: none;
+}
+
 .popup {
   padding: 16px 16px;
   backdrop-filter: blur(5px);
@@ -31,6 +38,7 @@
   right: 50px;
   top: 50px;
   z-index: 2;
+  cursor: pointer;
 }
 
 .pop-card {
